add cli and main file
